@@ -1,56 +1,61 @@
 {
-  "name": "openstack-infra/openstackid-resources",
-  "description": "The OpenStackId Resource Server.",
-  "keywords": [
-    "framework",
-    "laravel"
-  ],
-  "license": "MIT",
-  "type": "project",
-  "require": {
-    "laravel/framework": "5.0.*",
-    "predis/predis": "1.0.1",
-    "php": ">=5.4.0",
-    "guzzlehttp/guzzle": "5.3.0",
-    "ezyang/htmlpurifier": "4.7.0"
-  },
-  "require-dev": {
-    "phpunit/phpunit": "4.6.6",
-    "phpspec/phpspec": "~2.1",
-    "mockery/mockery": "0.9.4",
-    "squizlabs/php_codesniffer": "2.*",
-    "pragmarx/laravelcs": "*",
-    "glenscott/url-normalizer" : "1.4.0"
-  },
-  "autoload": {
-    "classmap": [
-      "database",
-      "app"
+    "name": "openstack-infra/openstackid-resources",
+    "description": "The OpenStackId Resource Server.",
+    "keywords": [
+      "framework",
+      "laravel"
     ],
-    "psr-4": {
-      "App\\": "app/"
+    "license": "MIT",
+    "type": "project",
+    "require": {
+        "php": ">=5.5.9",
+        "laravel/framework": "5.2.*",
+        "predis/predis": "1.0.1",
+        "guzzlehttp/guzzle": "5.3.0",
+        "ezyang/htmlpurifier": "4.7.0",
+        "glenscott/url-normalizer" : "^1.4",
+        "laravel-doctrine/orm":"1.1.*"
+    },
+    "require-dev": {
+        "fzaninotto/faker": "~1.4",
+        "mockery/mockery": "0.9.*",
+        "phpunit/phpunit": "~4.0",
+        "symfony/css-selector": "2.8.*|3.0.*",
+        "symfony/dom-crawler": "2.8.*|3.0.*"
+    },
+    "autoload": {
+        "classmap": [
+            "database",
+            "app",
+            "tests",
+            "Libs"
+        ],
+        "psr-4": {
+            "App\\": "app/"
+        }
+    },
+    "autoload-dev": {
+        "classmap": [
+            "tests/TestCase.php"
+        ]
+    },
+    "scripts": {
+        "post-root-package-install": [
+            "php -r \"copy('.env.example', '.env');\""
+        ],
+        "post-create-project-cmd": [
+            "php artisan key:generate"
+        ],
+        "post-install-cmd": [
+            "Illuminate\\Foundation\\ComposerScripts::postInstall",
+            "php artisan optimize"
+        ],
+        "post-update-cmd": [
+            "Illuminate\\Foundation\\ComposerScripts::postUpdate",
+            "php artisan optimize"
+        ]
+    },
+    "config": {
+        "preferred-install": "dist"
     }
-  },
-  "autoload-dev": {
-    "classmap": [
-      "tests"
-    ]
-  },
-  "scripts": {
-    "post-install-cmd": [
-      "php artisan clear-compiled",
-      "php artisan optimize"
-    ],
-    "post-update-cmd": [
-      "php artisan clear-compiled",
-      "php artisan optimize"
-    ],
-    "post-create-project-cmd": [
-      "php -r \"copy('.env.example', '.env');\"",
-      "php artisan key:generate"
-    ]
-  },
-  "config": {
-    "preferred-install": "dist"
-  }
 }
