@@ -1,37 +1,58 @@
-<?php namespace App\Http;
+<?php
+
+namespace App\Http;
 
 use Illuminate\Foundation\Http\Kernel as HttpKernel;
 
 class Kernel extends HttpKernel
 {
-
     /**
      * The application's global HTTP middleware stack.
+     *
+     * These middleware are run during every request to your application.
+     *
      * @var array
      */
     protected $middleware = [
-        'Illuminate\Foundation\Http\Middleware\CheckForMaintenanceMode',
-        'Illuminate\Cookie\Middleware\EncryptCookies',
-        'Illuminate\Cookie\Middleware\AddQueuedCookiesToResponse',
-        'Illuminate\Session\Middleware\StartSession',
-        'Illuminate\View\Middleware\ShareErrorsFromSession',
-        'App\Http\Middleware\CORSMiddleware',
-        'App\Http\Middleware\SecurityHTTPHeadersWriterMiddleware',
+        \Illuminate\Foundation\Http\Middleware\CheckForMaintenanceMode::class,
+        \App\Http\Middleware\EncryptCookies::class,
+        \Illuminate\Cookie\Middleware\AddQueuedCookiesToResponse::class,
+        \Illuminate\Session\Middleware\StartSession::class,
+        \Illuminate\View\Middleware\ShareErrorsFromSession::class,
+        \App\Http\Middleware\VerifyCsrfToken::class,
+        \App\Http\Middleware\CORSMiddleware::class,
+        \App\Http\Middleware\SecurityHTTPHeadersWriterMiddleware::class,
+    ];
+
+    /**
+     * The application's route middleware groups.
+     *
+     * @var array
+     */
+    protected $middlewareGroups = [
+        'web' => [
+        ],
+        'api' => [
+        ],
     ];
 
     /**
      * The application's route middleware.
+     *
+     * These middleware may be assigned to groups or used individually.
+     *
      * @var array
      */
     protected $routeMiddleware = [
-        'auth'             => 'App\Http\Middleware\Authenticate',
-        'auth.basic'       => 'Illuminate\Auth\Middleware\AuthenticateWithBasicAuth',
-        'guest'            => 'App\Http\Middleware\RedirectIfAuthenticated',
-        'oauth2.protected' => 'App\Http\Middleware\OAuth2BearerAccessTokenRequestValidator',
-        'rate.limit'       => 'App\Http\Middleware\RateLimitMiddleware',
-        'etags'            => 'App\Http\Middleware\ETagsMiddleware',
-        'cache'            => 'App\Http\Middleware\CacheMiddleware',
-        'ssl'              => 'App\Http\Middleware\SSLMiddleware',
+        'auth'             => \App\Http\Middleware\Authenticate::class,
+        'auth.basic'       => \Illuminate\Auth\Middleware\AuthenticateWithBasicAuth::class,
+        'can'              => \Illuminate\Foundation\Http\Middleware\Authorize::class,
+        'guest'            => \App\Http\Middleware\RedirectIfAuthenticated::class,
+        'throttle'         => \Illuminate\Routing\Middleware\ThrottleRequests::class,
+        'oauth2.protected' => \App\Http\Middleware\OAuth2BearerAccessTokenRequestValidator::class,
+        'rate.limit'       => \App\Http\Middleware\RateLimitMiddleware::class,
+        'etags'            => \App\Http\Middleware\ETagsMiddleware::class,
+        'cache'            => \App\Http\Middleware\CacheMiddleware::class,
+        'ssl'              => \App\Http\Middleware\SSLMiddleware::class,
     ];
-
 }
