@@ -1,4 +1,4 @@
-<?php
+<?php namespace models\utils;
 /**
  * Copyright 2015 OpenStack Foundation
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -12,42 +12,95 @@
  * limitations under the License.
  **/
 
-namespace models\utils;
-
+use Doctrine\ORM\Mapping AS ORM;
+use Doctrine\ORM\Query;
+use Doctrine\ORM\QueryBuilder;
+use Registry;
 /***
+ * @ORM\MappedSuperclass
  * Class SilverstripeBaseModel
  * @package models\utils
  */
-class SilverstripeBaseModel extends BaseModelEloquent implements IEntity
+class SilverstripeBaseModel implements IEntity
 {
-    protected $primaryKey ='ID';
-
-    protected $connection = 'ss';
-
-    protected $stiClassField = 'ClassName';
-
-    const CREATED_AT = 'Created';
-
-    const UPDATED_AT = 'LastEdited';
-
-    protected function isAllowedParent($parent_name)
+    /**
+     * @return mixed
+     */
+    public function getCreated()
     {
-        $res = parent::isAllowedParent($parent_name);
-        if(!$res) return false;
-        return !(str_contains($parent_name, 'SilverstripeBaseModel'));
-    }
-
-    public function __construct($attributes = array())
-    {
-        parent::__construct($attributes);
-        $this->ClassName = $this->table;
+        return $this->created;
     }
 
     /**
-     * @return int
+     * @param mixed $created
      */
+    public function setCreated($created)
+    {
+        $this->created = $created;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getLastEdited()
+    {
+        return $this->last_edited;
+    }
+
+    /**
+     * @param mixed $last_edited
+     */
+    public function setLastEdited($last_edited)
+    {
+        $this->last_edited = $last_edited;
+    }
+   /**
+     * @ORM\Id
+     * @ORM\Column(name="ID", type="integer", unique=true, nullable=false)
+     */
+    protected $id;
+
+    /**
+     * @ORM\Column(name="Created", type="datetime")
+     */
+    protected $created;
+
+    /**
+     * @ORM\Column(name="LastEdited", type="datetime")
+     */
+    protected $last_edited;
+
+    /**
+    * @return int
+    */
     public function getIdentifier()
     {
-        return (int)$this->ID;
+        return (int)$this->id;
+    }
+
+    public function getId(){
+        return $this->getIdentifier();
+    }
+
+    public function __construct()
+    {
+        $now               = new \DateTime();
+        $this->created     = $now;
+        $this->last_edited = $now;
+    }
+
+    /**
+     * @return QueryBuilder
+     */
+    public function createQueryBuilder(){
+        return Registry::getManager(self::ConnectionName)->createQueryBuilder();
+    }
+
+    /**
+     * @param string $dql
+     * @return Query
+     */
+    public function createQuery($dql){
+        return Registry::getManager(self::ConnectionName)->createQuery($dql);
     }
 }
