@@ -1,18 +1,19 @@
 <?php
 
-use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Seeder;
+use Illuminate\Database\Eloquent\Model;
 
 class DatabaseSeeder extends Seeder
 {
-
     /**
      * Run the database seeds.
+     *
      * @return void
      */
     public function run()
     {
-        Model::unguard();
+        // $this->call(UsersTableSeeder::class);
+	Model::unguard();
         $this->call('ApiSeeder');
         $this->call('ApiScopesSeeder');
         $this->call('ApiEndpointsSeeder');
